split llm agent tests into modules
